create RootCmd for teresa-server command line interface
